Update examples
* remove bootstrap
* fix an error with button icons
* convert babel to typescript
* clean out other unused dependencies
* add explanatory text to first example
* get rid of global jQuery in the embed-webpack.ts module
