+++ jupyter-js-widgets/examples/web/index.html
@@ -2,5 +2,8 @@
     <head>
         <script src="index.built.js"></script>
     </head>
-    <body />
+    <body>
+        <h1>Widget Example 1</h1>
+        <p>In this widget example, you will see a text box widget and a rendered HTML widget which are linked together. Type into the text box to see the HTML widget automatically updated.
+    </body>
 </html>
