+++ jupyter-js-widgets/examples/web2/widget_code.json
@@ -2,7 +2,7 @@
     "from ipywidgets import VBox, jsdlink, IntSlider, Button",
     "",
     "s1, s2 = IntSlider(max=200, value=100), IntSlider(value=40)",
-    "b = Button(icon='fa-legal')",
+    "b = Button(icon='legal')",
     "jsdlink((s1, 'value'), (s2, 'max'))",
     "VBox([s1, s2, b])"
 ]
