+++ jupyter-js-widgets/examples/web2/widget_state.json
@@ -41,7 +41,7 @@
         "model_module": "jupyter-js-widgets",
         "state": {
             "layout": "IPY_MODEL_f326c34c26034acdbab8ca4610d49bae",
-            "icon": "fa-legal"
+            "icon": "legal"
         },
         "views": []
     },
@@ -88,7 +88,7 @@
                         "collapsed": true
                     },
                     "cell_type": "code",
-                    "source": "from ipywidgets import VBox, jsdlink, IntSlider, Button\n\ns1, s2 = IntSlider(max=200, value=100), IntSlider(value=40)\nb = Button(icon='fa-legal')\njsdlink((s1, 'value'), (s2, 'max'))\nVBox([s1, s2, b])",
+                    "source": "from ipywidgets import VBox, jsdlink, IntSlider, Button\n\ns1, s2 = IntSlider(max=200, value=100), IntSlider(value=40)\nb = Button(icon='legal')\njsdlink((s1, 'value'), (s2, 'max'))\nVBox([s1, s2, b])",
                     "execution_count": 1,
                     "outputs": []
                 },
