+++ jupyter-js-widgets/examples/web3/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "prepublish": "npm run build",
     "build": "npm run build:es5 && npm run build:bundle",
-    "build:es5": "babel src --out-dir es5 --presets es2015",
+    "build:es5": "tsc --project src",
     "build:bundle": "webpack",
     "host": "http-server",
     "test": "npm run test:default",
@@ -15,25 +15,19 @@
   "author": "IPython",
   "license": "BSD-3-Clause",
   "dependencies": {
-    "bootstrap": "^3.3.6",
+    "@jupyterlab/services": "^0.24.0",
+    "@types/codemirror": "0.0.33",
     "codemirror": "^5.9.0",
     "font-awesome": "^4.5.0",
     "http-server": "^0.8.5",
-    "jquery": "^2.1.4",
-    "jquery-ui": "^1.10.5",
-    "jupyter-js-services": "0.10.2",
     "jupyter-js-widgets": "file:../.."
   },
   "devDependencies": {
-    "babel-cli": "^6.3.15",
-    "babel-core": "^6.10.4",
-    "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.3.13",
-    "bower": "^1.7.0",
     "css-loader": "^0.23.1",
     "file-loader": "^0.8.5",
     "json-loader": "^0.5.4",
     "style-loader": "^0.13.0",
+    "typescript": "^2.0.3",
     "url-loader": "^0.5.7",
     "webpack": "^1.12.10"
   }
