+++ jupyter-js-widgets/examples/web3/src/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "declaration": true,
+    //"noImplicitAny": true,
+    "lib": ["dom", "es5", "es2015.promise"],
+    "noEmitOnError": true,
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "target": "ES5",
+    "outDir": "../lib"
+  }
+}
