+++ jupyter-js-widgets/examples/web3/widget_code.json
@@ -5,9 +5,10 @@
     "s = IntSlider(max=200, value=100)",
     "t = Text()",
     "",
-    "def update_text(change):",
+    "def update_text(change=None):",
     "    t.value = str(s.value ** 2)",
     "",
     "s.observe(update_text, names='value')",
+    "update_text()",
     "display(VBox([s, t]))"
 ]
