+++ jupyter-js-widgets/examples/web4/index.html
@@ -57,7 +57,7 @@
                 "model_module": "jupyter-js-widgets",
                 "state": {
                     "layout": "IPY_MODEL_f326c34c26034acdbab8ca4610d49bae",
-                    "icon": "fa-legal"
+                    "icon": "legal"
                 },
                 "views": []
             },
