+++ jupyter-js-widgets/package.json
@@ -76,7 +76,6 @@
     "@types/backbone": "^1.3.33",
     "@types/semver": "^5.3.30",
     "backbone": "1.2.0",
-    "bootstrap": "^3.3.5",
     "d3-format": "^0.5.1",
     "font-awesome": "^4.5.0",
     "jquery": "^3.1.1",
